perf(teacher): ask for the profile directory in one constant-size request
The teacher view resolved authors by listing the ids on screen, which
PostgREST puts in the query string as one UUID each: 3KB today at 80 song
authors, and growing with every new class. PostgREST sits behind the usual
~8KB URL limit, so roughly two more schools and the lookup would just start
failing — a cliff, not a slowdown.

Teachers can already read every profile (migration 0011), so fetch them all
instead: 3048 bytes of URL becomes 97, and stays there no matter how far the
app spreads. It also drops the incremental-fetch bookkeeping the old hook
needed (the fetched-ids ref, the sorted-key dependency), since there is now
one request rather than one per new batch.

Emails are unaffected — that RPC passes ids in a POST body, which has no
comparable limit.

Not an Edge Requests concern either way: these are Supabase calls that never
touch Vercel. /teacher itself is one more page (HTML plus immutable chunks)
used by two teachers.

tsc/lint green, 117 tests; /teacher still renders empty for a non-teacher,
no console errors.

Refs #117
